Allow optional override of itemRenderer in AsyncComboField, ComboField and SelectField
