+++ cmp/form/SelectField.js
@@ -5,9 +5,8 @@
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
 
-import {isObject} from 'lodash';
 import {hoistComponent, elemFactory} from 'hoist/core';
-import {Classes, menuItem, select, button} from 'hoist/kit/blueprint';
+import {Classes, select, button} from 'hoist/kit/blueprint';
 
 import {HoistField} from './HoistField';
 
@@ -18,6 +17,7 @@ import {HoistField} from './HoistField';
  *
  * @prop options, collection of form [{value: object, label: string}, ...] or [val, val, ...]
  * @prop placeholder, text to display when control is empty
+ * @prop itemRenderer, optional custom itemRenderer, a function that receives (item, itemProps)
  */
 @hoistComponent()
 export class SelectField extends HoistField {
@@ -29,7 +29,7 @@ export class SelectField extends HoistField {
     delegateProps = ['className', 'style', 'disabled'];
 
     render() {
-        const {style, width, options, placeholder, disabled} = this.props;
+        const {style, width, options, placeholder, itemRenderer, disabled} = this.props;
 
         const value = this.renderValue;
 
@@ -37,13 +37,7 @@ export class SelectField extends HoistField {
             popoverProps: {popoverClassName: Classes.MINIMAL},
             $items: options,
             onItemSelect: this.onItemSelect,
-            itemRenderer: (item, itemProps) => {
-                let isObj = isObject(item) && item.value,
-                    value = isObj ? item.value : item,
-                    label = isObj ? item.label : item;
-                if (label === null) label = '-';
-                return menuItem({key: value, text: label.toString(), onClick: itemProps.handleClick});
-            },
+            itemRenderer: itemRenderer || this.defaultItemRenderer,
             filterable: false,
             item: button({
                 rightIcon: 'caret-down',
